requirements.txt
streamlit
openai
python-dotenv
pymysql
aiohttp